+++ raspiCamSrv/__init__.py
@@ -58,6 +58,7 @@ def create_app(test_config=None):
     #>>>>> Explicitely set specific log levels. Leave "werkzeug" at INFO
     logging.getLogger("werkzeug").setLevel(logging.INFO)
     #logging.getLogger("raspiCamSrv.camera_pi").setLevel(logging.DEBUG)
+    #logging.getLogger("raspiCamSrv.webcam").setLevel(logging.DEBUG)
     #logging.getLogger("raspiCamSrv.motionDetector").setLevel(logging.DEBUG)
     
     #>>>>> Set log level for picamera2 (DEBUG, INFO, WARNING, ERROR)